Working out cli UX
Currently, if a key (-k) is not defined per the API, a guaranteed GUID key is returned;
thus the below working example will create a new entry without specifying the key (only
the -v value); then it awk's out the key from the return string and uses that to call
`kv get -k` on the key. Neat!

  perl -I./lib/ bin/kv get -k $(perl -I./lib bin/kv set -v foobar | awk '{print $1}')

Assuming this is stored in plain text it is recommended that anyone using this encrypt
their values somehow. That is outside the scope of this module, though.